refactor: `github.strategy.ts` 파일 위치 변경
- `auth` 와 관련되며, 여러 `strategy` 가 생길 수 있으므로, `auth/strategy` 에 위치
