+++ backend/src/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
-import { GithubStrategy } from "../config/github.strategy";
+import { GithubStrategy } from "./strategy/github.strategy";
 import { UserRepository } from "../user/user.repository";
 
 @Module({
